Trait shenlong::core::sierra::libfunc::LibfuncProcessor

pub trait LibfuncProcessor<'a, 'ctx> {
    fn to_llvm(
        &self,
        module: &Module<'ctx>,
        context: &Context,
        builder: &Builder<'ctx>,
        func_name: &str
    ) -> Result<(), CompilerErr>;
}

LibfuncProcessor is a trait that will be implemented by all libfunc processors. It will be used to generate the LLVM IR for the libfunc.

Required Methods

fn to_llvm(
&self,
module: &Module<'ctx>,
context: &Context,
builder: &Builder<'ctx>,
func_name: &str
) -> Result<(), CompilerErr>

Generate the LLVM IR for the libfunc. The function will be added to the module. The function will be named fn_name.

Arguments

fn_name - The name of the function.

llvm_type - The type handled by the function, as input parameter or/and return type.

module - The module where the function will be added.

context - The context used to create the function.

builder - The builder used to create the function.