Module shenlong::core::sierra::llvm_compiler

LLVM compiler.

Description

This module contains the compiler that compiles a Sierra program to LLVM IR. The compilation can be triggered by calling different functions, depending on the input source. The input source can be a Sierra program file path, a Sierra program code as a string, or a Sierra program as a Program object. If the Sierra program is compiled from a file, the file will be read and the program code will be extracted. The program code will then be parsed into a Program object. The Program object will then be compiled to LLVM IR and written to a file. Here are the 3 functions that can be used to trigger the compilation:

compile_from_file: When the input source is a Sierra program file path.

compile_from_code: When the input source is a Sierra program code as a string.

compile_sierra_program_to_llvm: When the input source is a Sierra program as a Program object.

Compilation steps

The compilation process is split into multiple steps. The steps are executed in the following order:

Process the types.

Process the core library functions.

Process the statements.

Finalize the compilation.

State machine

The compiler is implemented as a state machine. The state machine is implemented using a HashMap that maps a CompilationStateTransition to a bool. The CompilationStateTransition is a tuple of two CompilationStates. The first state is the current state. The second state is the next state. The bool indicates whether the transition is valid. The state machine is initialized with all valid transitions. The state machine is updated after each compilation step. The state machine is used to ensure that the compilation steps are executed in the correct order. The state machine is also used to ensure that the compilation steps are executed only once.

Structs

Compiler

Compiler is the main entry point for the LLVM backend. It is responsible for compiling a Sierra program to LLVM IR.

Enums

CompilationState

Compilation state. This is used to keep track of the current compilation state. The reason is that the compilation process is split into multiple steps. The state will be used to implement a state machine that will keep track of the current compilation step. Goal is to ensure consistency in the order of the compilation steps. This is important because the compilation steps are not independent. For example, the type declarations must be processed before the statements. The state machine will ensure that the compilation steps are executed in the correct order. The state machine will also ensure that the compilation steps are executed only once.

Type Definitions

CompilationStateTransition

A compilation state transition. This is a tuple of two compilation states. The first state is the current state. The second state is the next state.